Restrukturert kode for bedre lesbarhet: - Pakket importering inn i div scope. - Renamet forretning => felles siden denne gjelder for både generisk og forretningsvalidering. - Div omstrukturering i generisk validering.
