+++ src/felles-komponenter/skjema/validering/felles.js
