+++ src/felles-komponenter/skjema/validering/index.js
@@ -1,7 +1,7 @@
-import { adresseKreves, norskPostNummer } from './adresse';
-import { minLengde, erPakrevet, kunTall, erDato, avhengerAvSann } from './generelt';
+import { adresseKreves, norskPostNummer } from './generisk/adresse';
+import { minLengde, erPakrevet, kunTall, erDato, avhengerAvSann } from './generisk';
 import { fulltNavn, fnr } from './generisk/person';
-import * as Forretning from './forretning';
+import * as Felles from './felles';
 
 export {
   adresseKreves,
@@ -13,5 +13,5 @@ export {
   kunTall,
   erDato,
   avhengerAvSann,
-  Forretning,
+  Felles,
 };
